feat : GetUserThumbnailController 분리
